# htdocs/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cooking Book API",
        "description": "REST API for Cooking Book application - manage recipes, comments, likes and categories",
        "contact": {
            "email": "admin@cookingbook.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/recipes": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Get list of recipes",
                "description": "Returns paginated list of recipes with optional search, filter, and sort",
                "operationId": "e6ebf01e62f938e7ce367c7b2fc94110",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in title and description",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter by category ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "difficulty",
                        "in": "query",
                        "description": "Filter by difficulty",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "easy",
                                "medium",
                                "hard"
                            ]
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort field",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "views_count",
                                "created_at",
                                "title",
                                "likes_count",
                                "comments_count"
                            ]
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of recipes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "description": "Recipe object",
                                                "type": "object"
                                            }
                                        },
                                        "metadata": {
                                            "properties": {
                                                "pagination": {
                                                    "properties": {
                                                        "current_page": {
                                                            "type": "integer"
                                                        },
                                                        "per_page": {
                                                            "type": "integer"
                                                        },
                                                        "total": {
                                                            "type": "integer"
                                                        },
                                                        "last_page": {
                                                            "type": "integer"
                                                        },
                                                        "from": {
                                                            "type": "integer",
                                                            "nullable": true
                                                        },
                                                        "to": {
                                                            "type": "integer",
                                                            "nullable": true
                                                        },
                                                        "has_more": {
                                                            "type": "boolean"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "next_cursor": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "filters": {
                                                    "properties": {
                                                        "search": {
                                                            "type": "string",
                                                            "nullable": true
                                                        },
                                                        "category_id": {
                                                            "type": "integer",
                                                            "nullable": true
                                                        },
                                                        "difficulty": {
                                                            "type": "string",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "sortings": {
                                                    "properties": {
                                                        "sort_by": {
                                                            "type": "string",
                                                            "nullable": true
                                                        },
                                                        "sort_order": {
                                                            "type": "string",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Create a new recipe",
                "description": "Use multipart/form-data to upload image. For ingredients array, use ingredients[0], ingredients[1], etc.",
                "operationId": "5ed3009a1c8ce5b784cac78533de2c73",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "instructions",
                                    "cooking_time",
                                    "servings",
                                    "difficulty"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Spaghetti Bolognese"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Classic Italian pasta dish"
                                    },
                                    "ingredients[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "instructions": {
                                        "type": "string",
                                        "example": "1. Cook pasta..."
                                    },
                                    "image": {
                                        "description": "Recipe image (jpeg,png,jpg,gif,webp, max 5MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "cooking_time": {
                                        "type": "integer",
                                        "example": 45
                                    },
                                    "servings": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "difficulty": {
                                        "type": "string",
                                        "enum": [
                                            "easy",
                                            "medium",
                                            "hard"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recipe created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/recipes/{id}": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Get recipe details",
                "operationId": "e59df102512a161f3e7be88aaa85a0e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recipe details"
                    },
                    "404": {
                        "description": "Recipe not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Update a recipe",
                "description": "Use POST with _method=PUT for multipart/form-data image upload",
                "operationId": "740ba44b40299d681616e23f8ed2effd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "ingredients[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "instructions": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "Recipe image (optional)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "cooking_time": {
                                        "type": "integer"
                                    },
                                    "servings": {
                                        "type": "integer"
                                    },
                                    "difficulty": {
                                        "type": "string",
                                        "enum": [
                                            "easy",
                                            "medium",
                                            "hard"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recipe updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - not owner or admin"
                    },
                    "404": {
                        "description": "Recipe not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Delete a recipe",
                "operationId": "a527903d2d7195e7456416e97f55faf2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recipe deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - not owner or admin"
                    },
                    "404": {
                        "description": "Recipe not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/recipes/{id}/increment-views": {
            "post": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Increment recipe view count",
                "operationId": "c0d585faa433581cc5a4865dc2005ccb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "View count incremented"
                    },
                    "404": {
                        "description": "Recipe not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}